requirements.txt
# Core dependencies for Minecraft AI agent
pygame>=2.5.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# macOS automation
pyautogui>=0.9.54
pynput>=1.7.6

# AI/ML frameworks
torch>=2.0.0
tensorflow>=2.13.0

# Screen capture and image processing
mss>=9.0.1
scikit-image>=0.21.0

# Configuration and utilities
pyyaml>=6.0
click>=8.1.0
python-dotenv>=1.0.0

# Development tools
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0